email-validation/requirements.txt
dnspython>=2.6.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
email-validator>=2.0.0
httpx>=0.27.0
